requirements.txt
anthropic==0.39.0
attr==0.3.2
backoff==2.2.1
chromadb==0.5.17
chunking_evaluation==0.1.0
jieba==0.42.1
numpy==2.1.3
openai==1.54.4
pdfplumber==0.11.4
rank_bm25==0.2.2
sentence_transformers==3.2.1
tiktoken==0.8.0
torch==2.4.1+cu121
tqdm==4.66.5
